Update mbpoll.c
Add support to write polled data into a CSV file. This implementation
honors the format setting. The CSV file will have a header line
generated containing the column names. Columns include the UTC timestamp
in RFC3339 format as well as the device IDs and register numbers. For
int16 values, the output differes from the console output: the columns
are duplicated and both the signed and unsigned values are recorded in
the CSV. This makes it easier to process the CSV on popular spreadsheet
applications without having to do a bunch of nasty conversions.